renaming header dashboard component
